services/text-detection/requirements.txt
# Core Deep Learning & NLP
torch
torchvision
transformers
accelerate
bitsandbytes
easyocr

# Data Handling & Scientific Computing
pandas
numpy
opencv-python-headless
Pillow
scikit-learn
PyYAML

# Metrics & Utilities
shapely
python-Levenshtein
nltk
natsort
gdown
tqdm
